Add support for disabling optimizing Eq.
We need to disable optimization on verilog code like "a == 1'b1". By
default, verilator optimizes the code to "a". However, since cases will
be translated to if-statements, removing the eq statment makes finite
state machine detection difficult.

In the current code, having "--skip-opt-verilog-eq" in the verilator
parameter would disable the optimization described above.